use Float for opacity and shadowOpacity (#69)

+++ Sources/CALayer.swift
@@ -55,7 +55,7 @@ open class CALayer {
     }
 
     public var isHidden = false
-    public var opacity: CGFloat = 1
+    public var opacity: Float = 1
     public var cornerRadius: CGFloat = 0
 
     // TODO: Implement these!
@@ -64,7 +64,7 @@ open class CALayer {
 
     public var shadowPath: CGRect?
     public var shadowColor: CGColor?
-    public var shadowOpacity: CGFloat = 0
+    public var shadowOpacity: Float = 0
     public var shadowOffset: CGSize = .zero
     public var shadowRadius: CGFloat = 0
